Refactor CleanUpService to separate file
